delete notification channels if reconnection failed
since the Pulsar binary protocol is multiplexed, when there's a
connection failure (ex: proxy rebooting), we can have multiple tasks
waiting for the reconnection at the same time. We do not want to trigger
multiple reconnections for the same address, so we use a connection
manager to handle that. When we request a connection to a broker, the
manager checks if it has an existing connection, and if it is valid. If
not, it will trigger a reconnection, and set its status to Connecting.
Whenever a task requests the same connection, while it is already
connecting, the task will add a oneshot channel, to get notified when it
is done.

Here is the bug: if reconnection failed, the initial task got the error,
but not the other tasks, and the status was kept at Connecting. So the
tasks never got the error, and any new reconnection would see that it is
connecting, and add its own notification channel, waiting indefinitely.

This change makes sure that on any reconnection error, all tasks get the
error